split aws feature granularly
