+++ src/utils/cert_manager/challenge/dns01/mod.rs
@@ -3,7 +3,7 @@ mod azure;
 mod cloudflare;
 mod gcloud;
 mod pebble;
-#[cfg(feature = "aws")]
+#[cfg(feature = "aws-route53")]
 mod route53;
 mod token;
 
@@ -12,7 +12,7 @@ pub use azure::{AzureDnsProvider, ServicePrincipal};
 pub use cloudflare::CloudflareDnsProvider;
 pub use gcloud::GoogleCloudDnsProvider;
 pub use pebble::PebbleDnsProvider;
-#[cfg(feature = "aws")]
+#[cfg(feature = "aws-route53")]
 pub use route53::AwsRoute53DnsProvider;
 
 use std::{sync::Arc, time::Duration};
